requirements.txt
gspread==5.4.0
numpy==1.23.2
oauth2client==4.1.3
pandas==1.4.3
pytz==2022.2.1
google-api-python-client==2.113.0